refactor(direct-control): extract session state selection
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves selectCiv7TunerState and its private role/name/id normalizer into src/session/state.ts while keeping the public facade export in index.ts.

Preserves role/name/id state selection and state-not-found error classification. This does not change endpoint discovery, config resolution, socket/session lifecycle, command execution, reconnect behavior, health polling, or runtime-status projection. Session config/socket/reconnect source ownership, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan for session state-selection helper definition

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
